gerrit_comments/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gerrit-comments"
version = "0.1.0"
description = "Extract and reply to Gerrit review comments - agent skill for Claude"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Claude", email = "claude@anthropic.com"}
]
keywords = ["gerrit", "code-review", "comments", "agent", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control",
]
dependencies = [
    "pygerrit2>=2.0.15",
    "requests>=2.25.0",
    "flask>=3.0.2",
    "flask-caching>=2.1.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
gerrit-comments = "gerrit_comments.cli:main"

[project.urls]
Homepage = "https://github.com/anthropics/claude-code"
Documentation = "https://github.com/anthropics/claude-code"

[tool.setuptools.packages.find]
include = ["gerrit_comments*"]

[tool.pytest.ini_options]
testpaths = ["gerrit_comments/tests"]
markers = [
    "integration: marks tests that require network access to real Gerrit",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict= (Python 3.10+)
]

[tool.ruff.lint.per-file-ignores]
"gerrit_comments/tests/*" = ["B011"]  # allow assert False in tests
